% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_SRE.R
\name{assign_block}
\alias{assign_block}
\title{function assigning treatment assignment for given summary of SRE}
\usage{
assign_block(block.sum, null.max = 10^4)
}
\description{
output n-length treatment assignment vector regarding SRE.
}
