% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_CRE.R
\name{assign_CRE}
\alias{assign_CRE}
\title{Generate matrix of complete randomization assignments
output: a matrix with nperm columns, each column gives one assignment}
\usage{
assign_CRE(n, m, nperm)
}
\description{
Generate matrix of complete randomization assignments
output: a matrix with nperm columns, each column gives one assignment
}
