% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS-package.R
\docType{package}
\name{CMRSS-package}
\alias{CMRSS}
\alias{CMRSS-package}
\title{CMRSS: Combining Multiple Rank Sum Statistics}
\description{
Randomization inference for quantiles of individual treatment effects using combined rank sum statistics. Supports both completely randomized experiments (CRE) and stratified randomized experiments (SRE). The optimization can be performed using either the open-source HiGHS solver or the commercial Gurobi solver.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bowers-illinois-edu.github.io/CMRSS/}
  \item \url{https://github.com/bowers-illinois-edu/CMRSS}
  \item Report bugs at \url{https://github.com/bowers-illinois-edu/CMRSS/issues}
}

}
\author{
\strong{Maintainer}: David Kim \email{davidk9@illinois.edu}

Authors:
\itemize{
  \item David Kim \email{davidk9@illinois.edu}
  \item Xinran Li \email{xinranli@uchicago.edu}
}

Other contributors:
\itemize{
  \item Jake Bowers \email{jwbowers@uchicago.edu} [contributor]
}

}
\keyword{internal}
