% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_CRE.R
\name{com_conf_quant_larger_cre}
\alias{com_conf_quant_larger_cre}
\title{Simultaneous lower bounds using multiple rank sum statistics in CRE}
\usage{
com_conf_quant_larger_cre(
  Z,
  Y,
  methods.list,
  nperm = 10^4,
  set = "treat",
  Z.perm = NULL,
  alpha = 0.05,
  tol = 10^(-3)
)
}
\arguments{
\item{Z}{An \eqn{n} dimensional treatment assignment vector.}

\item{Y}{An \eqn{n} dimensional observed outcome vector.}

\item{methods.list}{A list of lists specifies the choice of the multiple
rank sum test statistics. For example, list(name = "Wilcoxon") means the Wilcoxon
rank sum statistic, list(name = "Stephenson", s = 10) means
the Stephenson rank sum statistic with parameter s = 10, and
list(name = "Polynomial", r = 2, std = TRUE, scale = FALSE) means using polynomial
rank score with standarizing statistics without scaling.}

\item{nperm}{A positive integer representing the number of
permutations for approximating the randomization distribution of
the rank sum statistic.}

\item{set}{set of quantiles of interests. If "treat", generate prediction intervals
for effect quantiles among treated units. If "control", generate prediction intervals
for effect quantiles among control units. If "all", generate confidence intervals for
all effect quantiles.}

\item{Z.perm}{A \eqn{n \times nperm} matrix that specifies the
permutated assignments for approximating the null distribution of
the test statistic.}

\item{alpha}{A numerical object, where 1-alpha indicates the
confidence level.}

\item{tol}{A numerical object specifying the precision of the
obtained confidence intervals. For example, if tol = 10^(-3),
then the confidence limits are precise up to 3 digits.}
}
\value{
A vector specifying lower limits of prediction (confidence) intervals for
quantiles k = 1 ~ m (or n - m, or n).
}
\description{
Simultaneous lower bounds using multiple rank sum statistics in CRE
}
