% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_CRE.R
\name{com_conf_quant_larger_trt}
\alias{com_conf_quant_larger_trt}
\title{Helper function for Simultaneous Inference for multiple quantiles on CRE
using combined p-value.}
\usage{
com_conf_quant_larger_trt(
  Z,
  Y,
  methods.list = NULL,
  nperm = 10^4,
  k.vec = NULL,
  Z.perm = NULL,
  alpha = 0.05,
  tol = 10^(-3),
  ind.sort.treat = NULL
)
}
\description{
Output: lower limits of prediction intervals for (prespecified) quantiles across the treated units.
}
