% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sere_comb.ci.R
\name{com_null_dist_block}
\alias{com_null_dist_block}
\title{calculating null distribution on cmrss}
\usage{
com_null_dist_block(
  Z,
  block,
  methods.list.all,
  null.max = 10^5,
  Z.perm.all = NULL,
  ms.list
)
}
\description{
calculating null distribution on cmrss
}
