% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_matrix_block.R
\name{comb_matrix_block}
\alias{comb_matrix_block}
\title{Calculating all possible test statistic values for every strata}
\usage{
comb_matrix_block(Z, Y, block, c, methods.list.all)
}
\arguments{
\item{Z}{n-dimensional treatment assignment vector, for every units.}

\item{Y}{n-dimensional observed outcome vector, for every units.}

\item{block}{n-dimensional vector specifying stratum status of each units.}

\item{c}{A scalar specifying the bounded null hypothesis.}

\item{method.list.all}{list of every stratified rank sum statistics, where each element(list)s is method for every stratum.List of stratified rank sum statistic. The structure should be list of lists of lists.}
}
\description{
Generate a list of lists, where elements are all possible test statistic value for every possible null hypothesis in each stratum,
using user-specified score function.
}
