% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_CRE.R
\name{comb_p_val_cre}
\alias{comb_p_val_cre}
\title{Function for calculating valid minimum p-value in CRE
Calculate a valid p-value, based on multiple random sum statistics,
for testing the null that at most n-k treated units having effects > c,
null hypothesis H0: \eqn{\tau_{(k)} \leq c}, or \eqn{H0: \tau_{(k)} \geq c}, or
\eqn{H0: \tau_{(k)} = c}, where \eqn{\tau_{(k)}} denotes individual
treatment effect at rank k.}
\usage{
comb_p_val_cre(
  Z,
  Y,
  k,
  c,
  methods.list,
  Z.perm = NULL,
  nperm,
  stat.null.mult = NULL
)
}
\arguments{
\item{Z}{An \eqn{n} dimensional treatment assignment vector.}

\item{Y}{An \eqn{n} dimensional outcome vector.}

\item{k}{An integer between 1 and n specifying which quantile of
individual effect is of interest.}

\item{c}{A numerical object specifying the threshold for the null
hypothesis.}

\item{methods.list}{A list of lists specifies the choice of the multiple
rank sum test statistics. For example, list(name = "Wilcoxon") means the Wilcoxon
rank sum statistic, list(name = "Stephenson", s = 10) means
the Stephenson rank sum statistic with parameter s = 10, and
list(name = "Polynomial", r = 2, std = TRUE, scale = FALSE) means using polynomial
rank score with standarizing statistics without scaling.}

\item{Z.perm}{A \eqn{n \times nperm} matrix that specifies the
permutated assignments for approximating the null distribution of
the test statistic.}

\item{nperm}{A positive integer representing the number of
permutations for approximating the randomization distribution of
the rank sum statistic.}

\item{stat.null.mult}{A matrix whose empiricial distribution
approxmiates the randomization distribution of multiple rank statistics.}

\item{null.dist.comb}{}
}
\value{
Combined p-value for testing the specified null hypothesis of
  interest in CRE.
}
\description{
Function for calculating valid minimum p-value in CRE
Calculate a valid p-value, based on multiple random sum statistics,
for testing the null that at most n-k treated units having effects > c,
null hypothesis H0: \eqn{\tau_{(k)} \leq c}, or \eqn{H0: \tau_{(k)} \geq c}, or
\eqn{H0: \tau_{(k)} = c}, where \eqn{\tau_{(k)}} denotes individual
treatment effect at rank k.
}
