% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_min_stat.R
\name{combine_min_stat}
\alias{combine_min_stat}
\title{Combining minimum stat values in CRE}
\usage{
combine_min_stat(
  Z,
  Y,
  k,
  c,
  methods.list = NULL,
  ind.sort.treat = NULL,
  combine = max
)
}
\arguments{
\item{methods.list}{lists of methods. Should be list of lists}

\item{combine}{Function will be used for combining values. Maximum function is default.}
}
\description{
Combining minimum stat values in CRE
}
