% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_SRE.R
\name{Gurobi_sol_com}
\alias{Gurobi_sol_com}
\title{optimization function using integer programming by Gurobi
p is upper bound on the number of units with effects greater than c}
\usage{
Gurobi_sol_com(
  Z,
  block,
  weight,
  coeflists,
  p,
  ms_list,
  exact = TRUE,
  block.sum = NULL
)
}
\description{
optimization function using integer programming by Gurobi
p is upper bound on the number of units with effects greater than c
}
