% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_pval.R
\name{heatmap_pval}
\alias{heatmap_pval}
\title{heatmap_pval}
\usage{
heatmap_pval(
  Z,
  Y,
  k.vec,
  c.vec,
  block,
  methods.list.all,
  weight = NULL,
  stat.null = NULL,
  null.max = 10^5,
  Z.perm.all = NULL
)
}
\description{
A function to visualize multiple p-values corresponding to multiple choices of upper bounds and quantiles.
}
