% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_CRE.R
\name{min_p_multiple_rank_sum}
\alias{min_p_multiple_rank_sum}
\title{calculate the minimum p value from multiple rank sum statistics
output: minimum of p values calibrated by the distribution of minimum of tail probabilities}
\usage{
min_p_multiple_rank_sum(
  Z,
  Y,
  k,
  c,
  methods.list,
  Z.perm = NULL,
  nperm,
  stat.null.mult = NULL
)
}
\description{
calculate the minimum p value from multiple rank sum statistics
output: minimum of p values calibrated by the distribution of minimum of tail probabilities
}
