% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cre_comb_ci.R
\name{min_p_val_cre}
\alias{min_p_val_cre}
\title{function for calculating valid minimum p-value}
\usage{
min_p_val_cre(Z, Y, k, c, stat.null = NULL, methods.list, nperm)
}
\description{
function for calculating valid minimum p-value
}
