% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_SRE.R
\name{mu_sigma_list}
\alias{mu_sigma_list}
\title{function for calculating population mean / standard deviation for a single
rank sum statistics
XL: I change weight to a vector}
\usage{
mu_sigma_list(
  Z,
  block,
  weight,
  methods.list.all,
  scores.list.all = NULL,
  block.sum = NULL
)
}
\description{
function for calculating population mean / standard deviation for a single
rank sum statistics
XL: I change weight to a vector
}
