% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_SRE.R
\name{mu_sigma_single_weight_strat_rank_sum_stat}
\alias{mu_sigma_single_weight_strat_rank_sum_stat}
\title{function calculating mean and standard deviation
of each individual stratified rank sum statistic}
\usage{
mu_sigma_single_weight_strat_rank_sum_stat(
  Z,
  block,
  method.list.all,
  score.list.all = NULL,
  weight,
  block.sum = NULL
)
}
\description{
output is a list of mean and standard deviations of each rank sum statistic
}
