% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_CRE.R
\name{null_dist_multiple}
\alias{null_dist_multiple}
\title{Generate null distributions for multiple rank sum statistics}
\usage{
null_dist_multiple(
  n,
  m,
  methods.list = NULL,
  nperm = 10^5,
  Z.perm = NULL,
  chunk_size = 1000
)
}
\arguments{
\item{n}{Total number of units.}

\item{m}{Number of treated units.}

\item{methods.list}{A list of method specifications.}

\item{nperm}{Number of permutations.}

\item{Z.perm}{Optional pre-computed permutation matrix.}

\item{chunk_size}{Integer chunk size used to generate permutations when
\code{Z.perm} is \code{NULL}. Smaller values use less memory.}
}
\value{
An H x nperm matrix where H is the number of statistics.
}
\description{
Computes the randomization null distribution for multiple rank sum
statistics simultaneously under complete randomization.
}
\examples{
set.seed(1)
methods.list <- list(list(name = "Wilcoxon", scale = FALSE),
                     list(name = "Stephenson", s = 3, scale = FALSE))
dim(null_dist_multiple(n = 20, m = 12, methods.list = methods.list,
                       nperm = 500))

}
