% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_SRE.R
\name{poly_sd}
\alias{poly_sd}
\title{Standard deviation for polynomial rank transformation}
\usage{
poly_sd(r)
}
\arguments{
\item{r}{Power parameter for polynomial rank transformation.}
}
\value{
Standard deviation value.
}
\description{
Calculates the standard deviation of a polynomial rank score with
parameter r using the analytical formula.
}
\keyword{internal}
