% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_comb_auto.R
\name{pval_comb_auto}
\alias{pval_comb_auto}
\title{pval_comb_auto}
\usage{
pval_comb_auto(
  Z,
  Y,
  k,
  c,
  block,
  null.max = 10^3,
  STRA_CODE,
  statistic = FALSE,
  opt.method = "ILP_gurobi"
)
}
\description{
Function to calculate desired p-value combining 10 methods without specification
}
