% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cre_comb_ci.R
\name{pval_cre}
\alias{pval_cre}
\title{function calculating one p-value for individual methods(including polynomial scores)}
\usage{
pval_cre(
  Z,
  Y,
  k,
  c,
  method.list,
  score = NULL,
  stat.null = NULL,
  nperm = 10^3,
  Z.perm = NULL,
  ind.sort.treat = NULL
)
}
\description{
function calculating one p-value for individual methods(including polynomial scores)
}
