% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_SRE.R
\name{score_all_blocks}
\alias{score_all_blocks}
\title{score for each of the blocks under a given method.list.all}
\usage{
score_all_blocks(nb, method.list.all)
}
\description{
output is a list of score output for each blocks, when a method.list.all is given.
method.list.all is a list of lists
}
