% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_SRE.R
\name{score_stratum}
\alias{score_stratum}
\title{Compute stratum-specific scores for multiple methods}
\usage{
score_stratum(n, methods.list)
}
\arguments{
\item{n}{Number of units in the stratum.}

\item{methods.list}{A list of H method specifications for this stratum.}
}
\value{
A list of H score vectors.
}
\description{
Generates score vectors for a single stratum under multiple methods.
}
\keyword{internal}
