% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_SRE.R
\name{single_weight_strat_rank_sum_stat}
\alias{single_weight_strat_rank_sum_stat}
\title{function to calculate single stratified rank sum statistic with weights
output is a scalar.}
\usage{
single_weight_strat_rank_sum_stat(
  Z,
  Y,
  block,
  method.list.all,
  score.list.all = NULL,
  weight,
  block.sum = NULL
)
}
\description{
function to calculate single stratified rank sum statistic with weights
output is a scalar.
}
