% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_SRE.R
\name{weight_scheme}
\alias{weight_scheme}
\title{giving the weight under given scheme}
\usage{
weight_scheme(block.sum, weight.name = "asymp.opt")
}
\description{
output is a vector of giving weight for prespecified option of weights.
}
