% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_SRE.R
\name{assign_block}
\alias{assign_block}
\title{Generate block-randomized treatment assignments}
\usage{
assign_block(block.sum, null.max = 10^4)
}
\arguments{
\item{block.sum}{A block summary object from \code{summary_block}.}

\item{null.max}{Number of permutations to generate.}
}
\value{
An n x null.max matrix of permuted treatment assignments.
}
\description{
Generates permuted treatment assignments that respect the block structure
of a stratified randomized experiment.
}
\keyword{internal}
