% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_SRE.R
\name{com_block_conf_quant_larger}
\alias{com_block_conf_quant_larger}
\title{Simultaneous bound for confidence interval using combined rank sum statistic 
on stratified randomized experiment}
\usage{
com_block_conf_quant_larger(
  Z,
  Y,
  block,
  set = "all",
  methods.list.all = NULL,
  weight.name = "asymp.opt",
  opt.method = "ILP_gurobi",
  stat.null = NULL,
  null.max = 10^4,
  tol = 0.01,
  alpha = 0.1
)
}
\arguments{
\item{Z}{An \eqn{n} dimensional treatment assignment vector.}

\item{Y}{An \eqn{n} dimensional observed outcome vector.}

\item{block}{An \eqn{n} dimensional vector specifying block of each
units.}

\item{set}{set of quantiles of interests. If "treat", generate prediction intervals
for effect quantiles among treated units. If "control", generate prediction intervals
for effect quantiles among control units. If "all", generate confidence intervals for
all effect quantiles.}

\item{methods.list.all}{A list of lists of lists. Corresponds to the
.  method for each stratum for each different stratified rank sum 
.  statistic.}

\item{weight.name}{Weighting method to be implemented. If
"asymp.opt", asymptotically optimal scheme under a class of local 
alternatives is adjusted, if "dist.free", design-free scheme is 
adjusted.}

\item{opt.method}{Optimization method. If "ILP_gurobi", implement
integer linear programming. If "LP_gurobi", implement relaxed
linear programming.}

\item{stat.null}{An vector whose empirical distribution
approximates the randomization distribution of the combined 
stratified rank sum statistic.}

\item{null.max}{A positive integer representing the number of
permutations for approximating the randomization distribution of
the rank sum statistic.}

\item{tol}{A numerical object specifying the precision of the
obtained confidence intervals. For example, if tol = 10^(-3),
then the confidence limits are precise up to 3 digits.}

\item{alpha}{A numerical object, where 1-alpha indicates the
confidence level.}

\item{k}{An integer between 1 and n specifying which quantile of
individual effect is of interest.}

\item{c}{A numerical object specifying the threshold for the null
hypothesis.}
}
\value{
A vector specifying lower limits of prediction (confidence) intervals for
quantiles k = 1 ~ m (or n - m, or n).
}
\description{
null hypothesis H0: \eqn{\tau_{(k)} \leq c}, or \eqn{H0: \tau_{(k)} \geq c}, or
\eqn{H0: \tau_{(k)} = c}, where \eqn{\tau_{(k)}} denotes individual
treatment effect at rank k.
}
