% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sere_comb.ci.R
\name{com_block_conf_quant_larger}
\alias{com_block_conf_quant_larger}
\title{Simultaneous confidence interval using CMRSS on SERE}
\usage{
com_block_conf_quant_larger(
  Z,
  Y,
  block,
  quantiles = NULL,
  set = "all",
  methods.list.all = NULL,
  opt.method = "ILP_gurobi",
  stat.null = NULL,
  null.max = 10^4,
  alpha = 0.1
)
}
\description{
Simultaneous confidence interval using CMRSS on SERE
}
