% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_SRE.R
\name{com_block_conf_quant_larger_trt}
\alias{com_block_conf_quant_larger_trt}
\title{Helper function for Simultaneous Inference for multiple quantiles on CRE}
\usage{
com_block_conf_quant_larger_trt(
  Z,
  Y,
  block,
  k.vec = NULL,
  methods.list.all,
  weight.name = "asymp.opt",
  opt.method = "ILP_gurobi",
  stat.null = NULL,
  null.max = 10^4,
  tol = 0.01,
  alpha = 0.1
)
}
\description{
Output is a lower limit of prediction intervals for prespecified quantiles.
}
