% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sere_comb.ci.R
\name{com_block_conf_quant_larger_trt}
\alias{com_block_conf_quant_larger_trt}
\title{helper function}
\usage{
com_block_conf_quant_larger_trt(
  Z,
  Y,
  block,
  quantiles = NULL,
  methods.list.all = NULL,
  opt.method = "ILP_gurobi",
  stat.null = NULL,
  null.max = 10^4,
  tol = 0.01,
  alpha = 0.1
)
}
\description{
helper function
}
