% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cre_comb_ci.R
\name{com_conf_quant_larger_cre}
\alias{com_conf_quant_larger_cre}
\title{Simultaneous confidence interval using CMRSS for CRE}
\usage{
com_conf_quant_larger_cre(
  Z,
  Y,
  methods.list = NULL,
  stat.null = NULL,
  nperm = 10^4,
  set = "prev",
  Z.perm = NULL,
  k.vec = NULL,
  alpha = 0.05,
  tol = 10^(-3),
  ind.sort.treat = NULL
)
}
\description{
Getting Lower bound for every null c's
}
