% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cre_comb_ci.R
\name{com_conf_quant_larger_trt}
\alias{com_conf_quant_larger_trt}
\title{Helper function for Simultaneous Inference for multiple quantiles on CRE}
\usage{
com_conf_quant_larger_trt(
  Z,
  Y,
  methods.list = NULL,
  stat.null = NULL,
  nperm = 10^4,
  k.vec = NULL,
  Z.perm = NULL,
  alpha = 0.05,
  tol = 10^(-3),
  ind.sort.treat = NULL
)
}
\description{
Helper function for Simultaneous Inference for multiple quantiles on CRE
}
