% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_SRE.R
\name{com_null_dist_block}
\alias{com_null_dist_block}
\title{generating null distribution of maximum among 
(1) weighted, (2) normalized multiple stratified rank sum statistic}
\usage{
com_null_dist_block(
  Z,
  block,
  methods.list.all,
  scores.list.all = NULL,
  null.max = 10^4,
  weight,
  block.sum = NULL,
  Z.perm = NULL
)
}
\description{
output is a vector approximating distribution of null distribution of maximum among test statistics.
}
