% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_null_dist_block.R
\name{com_null_dist_block}
\alias{com_null_dist_block}
\title{Generate randomization distribution for combined stratified rank sum statistic.}
\usage{
com_null_dist_block(
  Z,
  block,
  methods.list.all,
  null.max = 10^5,
  Z.perm.all = NULL,
  mu_sigma_list
)
}
\arguments{
\item{Z}{n-dimension treatment assignment vector, for every units}

\item{block}{Length n vector specifies stratum status for every units}

\item{null.max}{Total size of randomization distribution.}

\item{Z.perm.all}{User-specified assignment matrix, for every element of null distribution.}

\item{mu_sigma_list}{mu_sigma_list List of population mean / variance of every stratified rank sum statistics. This will usually be a output from \emph{mu_sigma_list} function.}

\item{method.list.all}{list of every stratified rank sum statistics, where each element(list)s is method for every stratum.List of stratified rank sum statistic. The structure should be list of lists of lists.}
}
\description{
Generate randomization distribution for combined stratified rank sum statistic.
}
