% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_SRE.R
\name{comb_matrix_block}
\alias{comb_matrix_block}
\title{calculating score values for given multiple stratified rank sum statistics}
\usage{
comb_matrix_block(
  Z,
  Y,
  block,
  c,
  methods.list.all,
  scores.list.all = NULL,
  block.sum = NULL
)
}
\description{
output is a list with H elements, 
each of the H elements is also a list of length B, 
which is a 2*(1+mb) matrix containing the minimum value of the block-specific 
test statistic given 0:mb numbers of units with effects greater than c
}
