% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cre_comb_ci.R
\name{comb_null_dist_cre}
\alias{comb_null_dist_cre}
\title{function generating null distribution for minimum among tail probabilities}
\usage{
comb_null_dist_cre(n, m, methods.list, Z.perm = NULL, nperm = 10^4)
}
\description{
function generating null distribution for minimum among tail probabilities
}
