% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_CRE.R
\name{comb_null_dist_cre}
\alias{comb_null_dist_cre}
\title{null distribution for minimum p-values
output: the distribution of the minimum p-value, particularly Monte Carlo samples from \overline{F} in Theorem 1 of the paper}
\usage{
comb_null_dist_cre(
  n,
  m,
  methods.list,
  Z.perm = NULL,
  nperm = 10^4,
  stat.null.mult = NULL
)
}
\description{
null distribution for minimum p-values
output: the distribution of the minimum p-value, particularly Monte Carlo samples from \overline{F} in Theorem 1 of the paper
}
