% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{get_default_solver}
\alias{get_default_solver}
\title{Get default solver}
\usage{
get_default_solver()
}
\value{
Character string indicating the available solver
}
\description{
Returns the default solver based on availability. Prefers Gurobi if available
(to maintain backward compatibility with earlier versions/workflows), and
falls back to HiGHS.
}
\examples{
get_default_solver()
}
