% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gurobi_sol_com.R
\name{Gurobi_sol_com}
\alias{Gurobi_sol_com}
\title{Function implementing optimization for multiple stratified rank sum statistics, using Gurobi software.}
\usage{
Gurobi_sol_com(
  Z,
  block,
  weight = NULL,
  coeflists,
  p,
  mu_sigma_list,
  exact = TRUE
)
}
\arguments{
\item{Z}{n-dimension treatment assignment vector, for every units.}

\item{block}{n-dimension vector specifies stratum status of every units.}

\item{weight}{Vector which has same length with total number of stratum. Giving weight on each stratum before optimization. Basic option is number of treated units on each stratum.}

\item{coeflists}{All possible test statistics value, which will usually be a output from \emph{comb_matrix_block} function.}

\item{p}{N-k, where k is given in null hypothesis as \eqn{H_{N,k,c} = \tau_{k} \leq c}.}

\item{mu_sigma_list}{List of population mean / variance of every stratified rank sum statistics. This will usually be a output from \emph{mu_sigma_list} function.}

\item{exact}{## what was this??}
}
\description{
Obtain the value of test statistic by optimization, for \eqn{\underset{h}{max} \frac{t^{(h)}(Z, Y) - \mu_h }{\sigma_h}}.
}
