% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_CRE.R
\name{min_stat}
\alias{min_stat}
\title{calculating individual test statistic corresponds to Caughey et al.(2023)
output: minimum value of the rank sum statistic assuming n-k treated units can have infinite effects}
\usage{
min_stat(Z, Y, k, c, method.list = NULL, score = NULL, ind.sort.treat = NULL)
}
\description{
calculating individual test statistic corresponds to Caughey et al.(2023)
output: minimum value of the rank sum statistic assuming n-k treated units can have infinite effects
}
