% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_stat.R
\name{min_stat}
\alias{min_stat}
\title{Minimum stat value in H_{k,c}}
\usage{
min_stat(Z, Y, k, c, method.list = NULL, score = NULL, ind.sort.treat = NULL)
}
\arguments{
\item{Z}{Given assignment vector}

\item{Y}{Outcome vector}

\item{k}{k specified as kth largest treatment effect in the null hypothesis}

\item{c}{c specified as upper bound of kth largest treatment effect in the null hypothesis}

\item{method.list}{list of method.}

\item{score}{user-specified score vector, if available}

\item{ind.sort.treat}{user-specified alignment in treatment vector, if available}
}
\description{
The function
}
