% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_sigma_list.R
\name{mu_sigma_list}
\alias{mu_sigma_list}
\title{Function calculating population mean / variance for each stratified rank sum statistic,
to standardize test statistics globally before implementing optimization.}
\usage{
mu_sigma_list(Z, block, methods.list.all)
}
\arguments{
\item{Z}{n-dimensional treatment assignment vector, for every units.}

\item{block}{n-dimensional vector, which specifies stratum status of each units.}

\item{methods.list.all}{List of stratified rank sum statistic. The structure should be list of lists of lists.}
}
\description{
Function calculating population mean / variance for each stratified rank sum statistic,
to standardize test statistics globally before implementing optimization.
}
