% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_CRE.R
\name{null_dist}
\alias{null_dist}
\title{generating randomization null distribution of a single rank sum statistic
output: null dist of a single rank score statistics under the CRE}
\usage{
null_dist(n, m, method.list = NULL, score = NULL, nperm = 10^5, Z.perm = NULL)
}
\description{
generating randomization null distribution of a single rank sum statistic
output: null dist of a single rank score statistics under the CRE
}
