% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_CRE.R
\name{null_dist_multiple}
\alias{null_dist_multiple}
\title{null distribution of multiple rank sum statistics
output: matrix of null distributions from multiple rank statistics}
\usage{
null_dist_multiple(n, m, methods.list = NULL, nperm = 10^5, Z.perm = NULL)
}
\description{
null distribution of multiple rank sum statistics
output: matrix of null distributions from multiple rank statistics
}
