% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_comb.R
\name{pval_comb}
\alias{pval_comb}
\title{p_val_comb}
\usage{
pval_comb(
  Z,
  Y,
  k,
  c,
  block,
  methods.list.all,
  weight = NULL,
  stat.null = NULL,
  null.max = 10^5,
  Z.perm.all = NULL,
  statistic = TRUE,
  opt.method = "ILP_gurobi"
)
}
\description{
Calculate p-value for SCRE using combining multiple methods
}
