% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_SRE.R
\name{pval_comb_block}
\alias{pval_comb_block}
\title{Combined test statistic and combined p-value for randomization test
for quantiles of individual treatment effects}
\usage{
pval_comb_block(
  Z,
  Y,
  k,
  c,
  block,
  methods.list.all,
  weight.name = "asymp.opt",
  stat.null = NULL,
  null.max = 10^5,
  statistic = TRUE,
  opt.method = "ILP_gurobi"
)
}
\arguments{
\item{Z}{An \eqn{n} dimensional treatment assignment vector.}

\item{Y}{An \eqn{n} dimensional observed outcome vector.}

\item{k}{An integer between 1 and n specifying which quantile of
individual effect is of interest.}

\item{c}{A numerical object specifying the threshold for the null
hypothesis.}

\item{block}{An \eqn{n} dimensional vector specifying block of each
units.}

\item{methods.list.all}{A list of lists of lists. Corresponds to the
.  method for each stratum for each different stratified rank sum 
.  statistic.}

\item{weight.name}{Weighting method to be implemented. If
"asymp.opt", asymptotically optimal scheme under a class of local 
alternatives is adjusted, if "dist.free", design-free scheme is 
adjusted.}

\item{stat.null}{An vector whose empirical distribution
approximates the randomization distribution of the combined 
stratified rank sum statistic.}

\item{null.max}{A positive integer representing the number of
permutations for approximating the randomization distribution of
the rank sum statistic.}

\item{statistic}{logical; if TRUE (default), also prints the 
combined stratified rank sum statistic.}

\item{opt.method}{Optimization method. If "ILP_gurobi", implement
integer linear programming. If "LP_gurobi", implement relaxed
linear programming.}
}
\value{
The p-value (and test statistic) for testing the specified
  null hypothesis of interest.
}
\description{
Obtain the combined stratified rank sum statistic
and combined p-value for testing the given
null hypothesis H0: \eqn{\tau_{(k)} \leq c}, or \eqn{H0: \tau_{(k)} \geq c}, or
\eqn{H0: \tau_{(k)} = c}, where \eqn{\tau_{(k)}} denotes individual
treatment effect at rank k.
}
