% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_CRE.R
\name{pval_cre}
\alias{pval_cre}
\title{calculating p-value for single method
output: the p-value assuming at most n-k treated units having effects > c, using a single rank sum statistic}
\usage{
pval_cre(
  Z,
  Y,
  k,
  c,
  method.list,
  score = NULL,
  stat.null = NULL,
  nperm = 10^3,
  Z.perm = NULL,
  ind.sort.treat = NULL
)
}
\description{
calculating p-value for single method
output: the p-value assuming at most n-k treated units having effects > c, using a single rank sum statistic
}
