% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_score.R
\name{rank_score}
\alias{rank_score}
\title{Score function of the ranks}
\usage{
rank_score(
  n,
  method.list = list(name = "Polynomial", r, std = TRUE, scale = FALSE)
)
}
\arguments{
\item{n}{Sample size}

\item{method.list}{List of method. Standarizing and scaling is available for Polynomial rank score. Scaling is available for Wilcoxon and Stephenson.}
}
\description{
Calculating the scores of n units (ordered from low to high score) given specified test statistic, not allowing for ties.
}
