% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_CRE.R
\name{rank_score}
\alias{rank_score}
\title{Calculating score for given n
included polynomial rank score with Puri(1965)'s normalization (denoted by std)
output: rank scores of length n}
\usage{
rank_score(
  n,
  method.list = list(name = "Polynomial", r, std = TRUE, scale = FALSE)
)
}
\description{
Calculating score for given n
included polynomial rank score with Puri(1965)'s normalization (denoted by std)
output: rank scores of length n
}
