% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_SRE.R
\name{score_all_blocks}
\alias{score_all_blocks}
\title{Compute scores for all blocks}
\usage{
score_all_blocks(nb, method.list.all)
}
\arguments{
\item{nb}{Vector of block sizes.}

\item{method.list.all}{A list of method specifications, one per block.}
}
\value{
A list of score vectors, one per block.
}
\description{
Computes rank scores for each block according to block-specific methods.
}
\keyword{internal}
