% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{solver_available}
\alias{solver_available}
\title{Check if a solver is available}
\usage{
solver_available(solver)
}
\arguments{
\item{solver}{Character string: "gurobi" or "highs"}
}
\value{
Logical indicating if the solver is available
}
\description{
Check if a solver is available
}
\examples{
solver_available("highs")
solver_available("gurobi")
}
