% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_CRE.R
\name{sort_treat}
\alias{sort_treat}
\title{Sort treated units by outcome rank}
\usage{
sort_treat(Y, Z)
}
\arguments{
\item{Y}{An n-dimensional observed outcome vector.}

\item{Z}{An n-dimensional binary treatment assignment vector.}
}
\value{
Integer vector of indices of treated units, sorted by increasing outcome.
}
\description{
Returns the indices of treated units sorted by their outcome values
in increasing order.
}
\keyword{internal}
