% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRSS_CRE.R
\name{sort_treat}
\alias{sort_treat}
\title{sorting index among treated units, according to ranks of the outcome}
\usage{
sort_treat(Y, Z)
}
\description{
output: indices of treated units with increasing outcomes
}
